Bump required Bison version to 3.0.4
since that is the version I got after installing via Homebrew, and
building with that version resulted in several irrelevant source changes
to files in the gen/ directory.

I bumped the requirement so that subsequent changes are not made via an
older version, resulting in extraneous source changes. I.e., only allow
moving forward and not backward.