Add the beginnings of AST command validation.
The idea here is that we will keep two maps around, one which maps commands to
validation functions, and another which maps commands to usage strings. We will
need to do the same for validating tests so that we can turn the remaining
unit test failures green and merge to master.