Factor out user reporting to a separate API (issue #20).
This introduces a new class, "Diagnostic", which will handle the formatting
of all diagnostic issues. Right now, that is just a clang-style message of
the form:

Unrecognized test "bogus".
On line 1:
if bogus :contains :comparator "i;octet" "Subject" "MAKE MONEY FAST" { discard; }
   ^^^^^

This will be extended in the future to also support hints, as well as formatting
messages for other medium.