Allow test-lists to have a single element.
This fixes issue #47, which was a regression from attempting to allow
the convert command to be used as a test. We now simply add the convert
command to our test map to ensure it can be used that way.

Additionally, we make a change to allow laxer usage of parenthesis. This
is allowed by many (if not all) mail programs, since it is sometimes
more readable to encapsulate an entire test in parenthesis.