Adds a test-list AST object, and fixes validation for the `not` test
Fix for issue #48. Previously, we added tests in a vector which meant
that we couldn't distinguish between a single-item test-list and a
single test. This meant that for something like `not`, where the syntax
is:

    not <test>

we would incorrectly allow something like `if not (true) {...`, which
should rightly be flagged as incorrect.