Proton vendored: Fix issue with require deps check
This change fixes an issue across extensions requiring other extensions.
When a non-list format is used in `require`, prior functionality would
erroneously return false when checking for a dependency.

e.g.
```sieve
require "variables";
require "vnd.proton.eval";
```

This is because the prior lookup was searching the list of children of
the current `require` node.

To fix this, all lookups are patched with the existing
`requires_capability` function, which uses a `_capability_map` populated
by all `require` calls.